db/requirements.txt
asyncpg>=0.30.0
openai>=1.68.0
pgvector>=0.3.6
PyYAML>=6.0.2
opentelemetry-sdk>=1.30.0
opentelemetry-exporter-otlp-proto-http>=1.30.0
opentelemetry-instrumentation-httpx>=0.51b0
